+++ src/string.mli
@@ -79,9 +79,7 @@ val uncapitalize : t -> t
 *)
 
 (** [Caseless] compares and hashes strings ignoring case, so that for example
-    [Caseless.equal "OCaml" "ocaml"] and [Caseless.("apple" < "Banana")] are [true], and
-    [Caseless.Map], [Caseless.Table] lookup and [Caseless.Set] membership is
-    case-insensitive.
+    [Caseless.equal "OCaml" "ocaml"] and [Caseless.("apple" < "Banana")] are [true].
 
     [Caseless] also provides case-insensitive [is_suffix] and [is_prefix] functions, so
     that for example [Caseless.is_suffix "OCaml" ~suffix:"AmL"] and [Caseless.is_prefix
@@ -100,6 +98,13 @@ module Caseless : sig
 
   val is_suffix : t -> suffix:t -> bool
   val is_prefix : t -> prefix:t -> bool
+  val is_substring : t -> substring:t -> bool
+  val is_substring_at : t -> pos:int -> substring:t -> bool
+  val substr_index : ?pos:int -> t -> pattern:t -> int option
+  val substr_index_exn : ?pos:int -> t -> pattern:t -> int
+  val substr_index_all : t -> may_overlap:bool -> pattern:t -> int list
+  val substr_replace_first : ?pos:int -> t -> pattern:t -> with_:t -> t
+  val substr_replace_all : t -> pattern:t -> with_:t -> t
 end
 
 (** [index_exn] and [index_from_exn] raise [Caml.Not_found] or [Not_found_s] when [char]
@@ -133,7 +138,7 @@ module Search_pattern : sig
 
   (** [create pattern] preprocesses [pattern] as per KMP, building an [int array] of
       length [length pattern].  All inputs are valid. *)
-  val create : string -> t
+  val create : ?case_sensitive:bool (** default = true *) -> string -> t
 
   (** [matches pat str] returns true if [str] matches [pat] *)
   val matches : t -> string -> bool
@@ -164,6 +169,29 @@ module Search_pattern : sig
   val replace_first : ?pos:int -> t -> in_:string -> with_:string -> string
 
   val replace_all : t -> in_:string -> with_:string -> string
+
+  (**/**)
+
+  (*_ See the Jane Street Style Guide for an explanation of [Private] submodules:
+
+    https://opensource.janestreet.com/standards/#private-submodules *)
+  module Private : sig
+    type public = t
+
+    type t =
+      { pattern : string
+      ; case_sensitive : bool
+      ; kmp_array : int array
+      }
+    [@@deriving_inline equal, sexp_of]
+
+    val equal : t -> t -> bool
+    val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t
+
+    [@@@end]
+
+    val representation : public -> t
+  end
 end
 
 (** Substring search and replace convenience functions.  They call [Search_pattern.create]